Change: Renamed handler.py to verifier.py Feature: Defined and implemented a verification protocol for the server
The server reads in a JSON string to verify, and returns a JSON string reply.
The JSON consists of an object with two fields: status, and details.

When status is 'OK', it means that the verifier ran successfully, and that details is a list of strings. Each string is an error message describing an issue that the verifier found with the json provided in the message. If there were no issue, details is empty list.

When status is 'BAD_JSON', it means that the message wasn't a JSON string, and details is an error string from the JSON parser.

When status is 'VERIFIER_ERROR', it means that the verifier didn't return a list of strings ー either it threw an exception, or it returned some python object that wasn't a list of strings. In either case, details is an error string describing the error message.